+++ src/main/java/com/georgster/util/commands/wizard/input/ReactionListener.java
@@ -7,9 +7,19 @@ import discord4j.core.event.domain.message.ReactionAddEvent;
 import discord4j.core.event.domain.message.ReactionRemoveEvent;
 import discord4j.core.object.reaction.ReactionEmoji;
 
+/**
+ * Sends a message with reaction emojis and records when one is presented.
+ * The name of the emoji is recorded as the response.
+ */
 public class ReactionListener extends InputListener {
     private ReactionEmoji[] emojis;
 
+    /**
+     * Creates a new {@code ReactionListener} with the given parameters.
+     * 
+     * @param event The event that prompted this listener's creation.
+     * @param title The title to place onto messages.
+     */
     public ReactionListener(CommandExecutionEvent event, String title, ReactionEmoji... emojis) {
         super (event, title, "end");
         this.emojis = emojis;
@@ -17,6 +27,9 @@ public class ReactionListener extends InputListener {
         mustMatch(false, false);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public WizardState prompt(WizardState inputState) {
         for (ReactionEmoji emoji : emojis) {
             message.addReaction(emoji).block();
@@ -43,10 +56,7 @@ public class ReactionListener extends InputListener {
                     }
                 }
                 return false;
-            }).subscribe(event -> {
-                event.getEmoji().asEmojiData().name().ifPresent(this::setResponse);
-                event.getEmoji().asEmojiData().name().ifPresent(System.out::println);
-            })
+            }).subscribe(event -> event.getEmoji().asEmojiData().name().ifPresent(this::setResponse))
         );
 
         return waitForResponse(inputState);